+++ DX_Renderer/Core/Components/Vertices/IndexBuffer.cpp
@@ -50,7 +50,7 @@ namespace DXR
 		D3D12_INDEX_BUFFER_VIEW index_buffer_descriptor = {};
 		index_buffer_descriptor.BufferLocation = (*this->m_index_buffer)->GetGPUVirtualAddress();
 		index_buffer_descriptor.Format = DXGI_FORMAT_R32_UINT;
-		index_buffer_descriptor.SizeInBytes = this->m_indices.size() * sizeof(UINT);
+		index_buffer_descriptor.SizeInBytes = (UINT)(this->m_indices.size() * sizeof(UINT));
 
 		this->m_index_buffer_descriptor = index_buffer_descriptor;
 	}
